Add a cast from char to int before autoboxing
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@1042150 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/fonts/Font.java
@@ -178,9 +178,9 @@ public class Font {
      * @return the distance to adjust for kerning, 0 if there's no kerning
      */
     public int getKernValue(char ch1, char ch2) {
-        Map<Integer, Integer> kernPair = getKerning().get(ch1);
+        Map<Integer, Integer> kernPair = getKerning().get((int) ch1);
         if (kernPair != null) {
-            Integer width = kernPair.get(ch2);
+            Integer width = kernPair.get((int) ch2);
             if (width != null) {
                 return width.intValue() * getFontSize() / 1000;
             }
